+++ types/protocol.d.ts
@@ -3592,7 +3592,7 @@ export namespace Protocol {
 
         export type MixedContentResolutionStatus = ('MixedContentBlocked' | 'MixedContentAutomaticallyUpgraded' | 'MixedContentWarning');
 
-        export type MixedContentResourceType = ('AttributionSrc' | 'Audio' | 'Beacon' | 'CSPReport' | 'Download' | 'EventSource' | 'Favicon' | 'Font' | 'Form' | 'Frame' | 'Image' | 'Import' | 'JSON' | 'Manifest' | 'Ping' | 'PluginData' | 'PluginResource' | 'Prefetch' | 'Resource' | 'Script' | 'ServiceWorker' | 'SharedWorker' | 'SpeculationRules' | 'Stylesheet' | 'Track' | 'Video' | 'Worker' | 'XMLHttpRequest' | 'XSLT');
+        export type MixedContentResourceType = ('Audio' | 'Beacon' | 'CSPReport' | 'Download' | 'EventSource' | 'Favicon' | 'Font' | 'Form' | 'Frame' | 'Image' | 'Import' | 'JSON' | 'Manifest' | 'Ping' | 'PluginData' | 'PluginResource' | 'Prefetch' | 'Resource' | 'Script' | 'ServiceWorker' | 'SharedWorker' | 'SpeculationRules' | 'Stylesheet' | 'Track' | 'Video' | 'Worker' | 'XMLHttpRequest' | 'XSLT');
 
         export interface MixedContentIssueDetails {
             /**
@@ -4761,11 +4761,6 @@ export namespace Protocol {
             buckets: Bucket[];
         }
 
-        /**
-         * @experimental
-         */
-        export type PrivacySandboxAPI = ('BiddingAndAuctionServices' | 'TrustedKeyValue');
-
         export interface SetPermissionRequest {
             /**
              * Descriptor of permission to override.
@@ -4995,17 +4990,6 @@ export namespace Protocol {
 
         export interface AddPrivacySandboxEnrollmentOverrideRequest {
             url: string;
-        }
-
-        export interface AddPrivacySandboxCoordinatorKeyConfigRequest {
-            api: PrivacySandboxAPI;
-            coordinatorOrigin: string;
-            keyConfig: string;
-            /**
-             * BrowserContext to perform the action in. When omitted, default browser
-             * context is used.
-             */
-            browserContextId?: BrowserContextID;
         }
 
         /**
@@ -20015,29 +19999,9 @@ export namespace Protocol {
         }
 
         /**
-         * Protected audience interest group auction identifier.
-         */
-        export type InterestGroupAuctionId = string;
-
-        /**
-         * Enum of interest group access types.
-         */
-        export type InterestGroupAccessType = ('join' | 'leave' | 'update' | 'loaded' | 'bid' | 'win' | 'additionalBid' | 'additionalBidWin' | 'topLevelBid' | 'topLevelAdditionalBid' | 'clear');
-
-        /**
-         * Enum of auction events.
-         */
-        export type InterestGroupAuctionEventType = ('started' | 'configResolved');
-
-        /**
-         * Enum of network fetches auctions can do.
-         */
-        export type InterestGroupAuctionFetchType = ('bidderJs' | 'bidderWasm' | 'sellerJs' | 'bidderTrustedSignals' | 'sellerTrustedSignals');
-
-        /**
          * Enum of shared storage access scopes.
          */
-        export type SharedStorageAccessScope = ('window' | 'sharedStorageWorklet' | 'protectedAudienceWorklet' | 'header');
+        export type SharedStorageAccessScope = ('window' | 'sharedStorageWorklet' | 'header');
 
         /**
          * Enum of shared storage access methods.
@@ -20450,29 +20414,6 @@ export namespace Protocol {
             didDeleteTokens: boolean;
         }
 
-        export interface GetInterestGroupDetailsRequest {
-            ownerOrigin: string;
-            name: string;
-        }
-
-        export interface GetInterestGroupDetailsResponse {
-            /**
-             * This largely corresponds to:
-             * https://wicg.github.io/turtledove/#dictdef-generatebidinterestgroup
-             * but has absolute expirationTime instead of relative lifetimeMs and
-             * also adds joiningOrigin.
-             */
-            details: any;
-        }
-
-        export interface SetInterestGroupTrackingRequest {
-            enable: boolean;
-        }
-
-        export interface SetInterestGroupAuctionTrackingRequest {
-            enable: boolean;
-        }
-
         export interface GetSharedStorageMetadataRequest {
             ownerOrigin: string;
         }
@@ -20532,12 +20473,6 @@ export namespace Protocol {
 
         export interface GetRelatedWebsiteSetsResponse {
             sets: RelatedWebsiteSet[];
-        }
-
-        export interface SetProtectedAudienceKAnonymityRequest {
-            owner: string;
-            name: string;
-            hashes: string[];
         }
 
         /**
@@ -20622,66 +20557,6 @@ export namespace Protocol {
              * Storage bucket to update.
              */
             bucketId: string;
-        }
-
-        /**
-         * One of the interest groups was accessed. Note that these events are global
-         * to all targets sharing an interest group store.
-         */
-        export interface InterestGroupAccessedEvent {
-            accessTime: Network.TimeSinceEpoch;
-            type: InterestGroupAccessType;
-            ownerOrigin: string;
-            name: string;
-            /**
-             * For topLevelBid/topLevelAdditionalBid, and when appropriate,
-             * win and additionalBidWin
-             */
-            componentSellerOrigin?: string;
-            /**
-             * For bid or somethingBid event, if done locally and not on a server.
-             */
-            bid?: number;
-            bidCurrency?: string;
-            /**
-             * For non-global events --- links to interestGroupAuctionEvent
-             */
-            uniqueAuctionId?: InterestGroupAuctionId;
-        }
-
-        /**
-         * An auction involving interest groups is taking place. These events are
-         * target-specific.
-         */
-        export interface InterestGroupAuctionEventOccurredEvent {
-            eventTime: Network.TimeSinceEpoch;
-            type: InterestGroupAuctionEventType;
-            uniqueAuctionId: InterestGroupAuctionId;
-            /**
-             * Set for child auctions.
-             */
-            parentAuctionId?: InterestGroupAuctionId;
-            /**
-             * Set for started and configResolved
-             */
-            auctionConfig?: any;
-        }
-
-        /**
-         * Specifies which auctions a particular network fetch may be related to, and
-         * in what role. Note that it is not ordered with respect to
-         * Network.requestWillBeSent (but will happen before loadingFinished
-         * loadingFailed).
-         */
-        export interface InterestGroupAuctionNetworkRequestCreatedEvent {
-            type: InterestGroupAuctionFetchType;
-            requestId: Network.RequestId;
-            /**
-             * This is the set of the auctions using the worklet that issued this
-             * request.  In the case of trusted signals, it's possible that only some of
-             * them actually care about the keys being queried.
-             */
-            auctions: InterestGroupAuctionId[];
         }
 
         /**